Show eval messages if debugging
This change checks the fabrik and joomla debug functions and issues eval
with or without the @ symbol which prevents php messages being displayed
depending on the debug setting.

This change to calc.php is intended as a proof of concept, with the same
code being rolled out to other uses of eval once this has been agreed.