Update php.php
If the php validation script eval return false, It's not inevitably because the script fails but because the validation script return false. So we don't have to raise a script error.
We'd have to find a solution to distinguish if the eval function return false because of the script php or because of a syntax error.