update string.php line 242 (<= instead of <)
Used in list view with icon replacement feature. For example, if the value is "Création", the ord function in line 244 find 195 for the "é" thus striping it out as well as missing the last car. The result is "Cratio". The "replace umlauts" should not be necessary with the iconv function. Also the $fromEnc shouldn't be set dynamically according to the current localization ? Currently it miss the "é" to "e" replacement....