Allow calling fireEvent() on an element without event object
Some event callbacks require an event object to function. When calling
fireEvent() directly from JS (when writing JS for a fabrik element for
example) there's no event object available, so relax the callbacks to accept
a missing event object to make it easier for the user.

An alternative is to have the user pass in an event object to all fireEvent()
calls, like this:

$('object_name').fireEvent('change', {stop: function(){}})

But that's more complex than necessary and easy for a user to forget.