Update fabrik-element-calc-form.php
Calc element had been written to use a class that had been defined as a static "input-medium" size and it would ignore the columns (width) param.  This update translates cols into bootstrap class definition to allow variable width based on cols.  Also cleaned up the code to use heredoc for the HTML.

Cons: Previous, cols were ignored in bootstrap templates.  Now, it will have an effect. A dev might need to adjust their cols param to suit and it's not obvious or documented  that the translation points are 15, 40, and 60.  If this commit is accepted, notify me and I'll update the wiki  - bruced@zumasys.com